pyproject.toml
[project]
name = "teaming24"
version = "0.1.0"
description = "Multi-agent collaboration platform with Agentic Node (AN) support"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Teaming24 Team" }
]
keywords = ["ai", "agents", "multi-agent", "collaboration", "x402", "sandbox"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    # Core server
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "wsproto>=1.2.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",

    # x402 Payment Protocol (https://x402.gitbook.io/x402)
    "x402[evm]>=2.0.0",
    "web3>=7.0.0",
    "eth-account>=0.13.0",

    # Sandbox runtime — browser automation
    "playwright>=1.57.0",

    # Multi-Agent Framework
    "crewai>=1.0.0",

    # LLM routing (ANRouter)
    "litellm>=1.0.0",

    # OpenHands runtime integration
    "openhands-sdk>=1.11.0",
    "openhands-tools>=1.11.0",
    "openhands-workspace>=1.11.0",
]

[project.optional-dependencies]
# System metrics (optional — used by sandbox metrics collector)
metrics = ["psutil>=5.9.0"]

# Extra LLM provider SDKs (uncomment / install as needed)
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.8.0"]
langchain = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.0",
]

# Everything
all = ["teaming24[metrics,openai,anthropic,langchain]"]

[project.scripts]
teaming24 = "teaming24.server.cli:main"

[project.urls]
Homepage = "https://github.com/teaming24/teaming24"
Documentation = "https://teaming24.dev/docs"
Repository = "https://github.com/teaming24/teaming24"

# ---------------------------------------------------------------------------
# Build system
# ---------------------------------------------------------------------------

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["teaming24"]

# ---------------------------------------------------------------------------
# uv — primary package manager
# ---------------------------------------------------------------------------
# Quick start:
#   uv sync                      # install all deps
#   uv sync --group dev          # install with dev tools
#   uv run python main.py        # run without activating venv
#   uv run teaming24 --reload    # run via entry point
# ---------------------------------------------------------------------------

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "mypy>=1.0.0",
    "pre-commit>=4.0.0",
]

# ---------------------------------------------------------------------------
# Ruff — linter & formatter
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["teaming24"]

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=teaming24 --cov-report=term-missing"
